add support for dummy utxos in psbts
